+++ sound-mind-studio/tests/test_filter_configuration_panel.h
@@ -17,4 +17,7 @@ private slots:
     void changingDirectionalBlurLengthAndAngleUpdateConfigAndEmit();
     void changingSharpenAmountUpdatesConfigAndEmits();
     void setFilterConfigurationSyncsTheTypeComboAndNewSpinBoxesWithoutEmitting();
+    void selectingToneCurveShowsItsOwnGroup();
+    void editingTheToneCurveEditorUpdatesConfigAndEmits();
+    void setFilterConfigurationSyncsTheToneCurveEditorWithoutEmitting();
 };